only enterData
