--- src/platform/testing/e2e/cypress/support/commands/findData.js
@@ -1,30 +1,0 @@
-import { NO_LOG_OPTION } from '../index';
-import get from 'platform/utilities/data/get';
-
-/**
- * Looks up data for a field.
- * @param {Field}
- * @returns {*} Resolves to the field data if found or undefined otherwise.
- */
-Cypress.Commands.add('findData', field => {
-  let resolvedDataPath;
-
-  cy.get('@testData', NO_LOG_OPTION).then(testData => {
-    const relativeDataPath = field.key
-      .replace(/^root_/, '')
-      .replace(/_/g, '.')
-      .replace(/\._(\d+)\./g, (_, number) => `[${number}]`);
-
-    // Prefix the path to the array item if this field belongs to one.
-    resolvedDataPath = field.arrayItemPath
-      ? `${field.arrayItemPath}.${relativeDataPath}`
-      : relativeDataPath;
-
-    cy.wrap(get(resolvedDataPath, testData), NO_LOG_OPTION);
-  });
-
-  Cypress.log({
-    message: field.key,
-    consoleProps: () => ({ ...field, resolvedDataPath }),
-  });
-});
